+++ aws/resource_aws_acm_certificate_test.go
@@ -51,7 +51,8 @@ func TestAccAwsAcmResource_emailValidation(t *testing.T) {
 	})
 
 }
-func TestAccAwsAcmResource_certificateIssuingFlow(t *testing.T) {
+
+func TestAccAwsAcmResource_dnsValidationAndTagging(t *testing.T) {
 	if os.Getenv("ACM_CERTIFICATE_ROOT_DOMAIN") == "" {
 		t.Skip("Environment variable ACM_CERTIFICATE_ROOT_DOMAIN is not set")
 	}
@@ -96,23 +97,6 @@ func TestAccAwsAcmResource_certificateIssuingFlow(t *testing.T) {
 					resource.TestCheckResourceAttr("aws_acm_certificate.cert", "tags.Foo", "Baz"),
 				),
 			},
-			// Test that validation times out if certificate can't be validated
-			resource.TestStep{
-				Config:      testAccAcmCertificateWithValidationConfig(domain, sanDomain),
-				ExpectError: regexp.MustCompile("Expected certificate to be issued but was in state PENDING_VALIDATION"),
-			},
-			// Test that validation fails if given validation_fqdns don't match
-			resource.TestStep{
-				Config:      testAccAcmCertificateWithValidationConfigAndWrongFQDN(domain, sanDomain),
-				ExpectError: regexp.MustCompile("Certificate needs .* to be set but only .* was passed to validation_record_fqdns"),
-			},
-			// Test that validation succeeds once we provide the right DNS validation records
-			resource.TestStep{
-				Config: testAccAcmCertificateWithValidationAndRecordsConfig(root_zone_domain, domain, sanDomain),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestMatchResourceAttr("aws_acm_certificate_validation.cert", "certificate_arn", certificateArnRegex),
-				),
-			},
 			resource.TestStep{
 				ResourceName:      "aws_acm_certificate.cert",
 				ImportState:       true,
@@ -145,94 +129,6 @@ resource "aws_acm_certificate" "cert" {
   }
 }
 `, domain, sanDomain, tag1Key, tag1Value, tag2Key, tag2Value)
-}
-
-func testAccAcmCertificateWithValidationConfig(domain string, sanDomain string) string {
-	return fmt.Sprintf(`
-resource "aws_acm_certificate" "cert" {
-  domain_name = "%s"
-  validation_method = "DNS"
-  subject_alternative_names = ["%s"]
-
-  tags {
-    "Environment" = "Test"
-    "Foo" = "Baz"
-  }
-}
-
-
-resource "aws_acm_certificate_validation" "cert" {
-  certificate_arn = "${aws_acm_certificate.cert.arn}"
-  timeout = "20s"
-}
-`, domain, sanDomain)
-}
-
-func testAccAcmCertificateWithValidationConfigAndWrongFQDN(domain string, sanDomain string) string {
-	return fmt.Sprintf(`
-resource "aws_acm_certificate" "cert" {
-  domain_name = "%s"
-  validation_method = "DNS"
-  subject_alternative_names = ["%s"]
-
-  tags {
-    "Environment" = "Test"
-    "Foo" = "Baz"
-  }
-}
-
-
-resource "aws_acm_certificate_validation" "cert" {
-  certificate_arn = "${aws_acm_certificate.cert.arn}"
-  validation_record_fqdns = ["some-wrong-fqdn.example.com"]
-  timeout = "20s"
-}
-`, domain, sanDomain)
-}
-
-func testAccAcmCertificateWithValidationAndRecordsConfig(rootZoneDomain string, domain string, sanDomain string) string {
-	return fmt.Sprintf(`
-resource "aws_acm_certificate" "cert" {
-  domain_name = "%s"
-  validation_method = "DNS"
-  subject_alternative_names = ["%s"]
-
-  tags {
-    "Environment" = "Test"
-    "Foo" = "Baz"
-  }
-}
-
-
-data "aws_route53_zone" "zone" {
-  name = "%s."
-  private_zone = false
-}
-
-resource "aws_route53_record" "cert_validation" {
-  name = "${aws_acm_certificate.cert.domain_validation_options.0.resource_record_name}"
-  type = "${aws_acm_certificate.cert.domain_validation_options.0.resource_record_type}"
-  zone_id = "${data.aws_route53_zone.zone.id}"
-  records = ["${aws_acm_certificate.cert.domain_validation_options.0.resource_record_value}"]
-  ttl = 60
-}
-
-resource "aws_route53_record" "cert_validation_san" {
-  name = "${aws_acm_certificate.cert.domain_validation_options.1.resource_record_name}"
-  type = "${aws_acm_certificate.cert.domain_validation_options.1.resource_record_type}"
-  zone_id = "${data.aws_route53_zone.zone.id}"
-  records = ["${aws_acm_certificate.cert.domain_validation_options.1.resource_record_value}"]
-  ttl = 60
-}
-
-resource "aws_acm_certificate_validation" "cert" {
-  certificate_arn = "${aws_acm_certificate.cert.arn}"
-  validation_record_fqdns = [
-	"${aws_route53_record.cert_validation.fqdn}",
-	"${aws_route53_record.cert_validation_san.fqdn}"
-  ]
-}
-`, domain, sanDomain, rootZoneDomain)
 }
 
 func testAccCheckAcmCertificateDestroy(s *terraform.State) error {
